refactor: 타입스크립트 필요한 부분만 변경
