+++ src/assets/images.d.ts
@@ -1,0 +1,25 @@
+declare module '*.png' {
+  const value: string
+  export default value
+}
+
+declare module '*.jpg' {
+  const value: string
+  export default value
+}
+
+declare module '*.jpeg' {
+  const value: string
+  export default value
+}
+
+declare module '*.gif' {
+  const value: string
+  export default value
+}
+
+declare module '*.svg' {
+  const value: string
+  export default value
+}
+// 문자열로 타입선언하는 이유는 <img src="경로"> 이기 때문
