+++ src/types/product.ts
@@ -1,0 +1,13 @@
+export interface Product {
+  createdAt: string
+  favoriteCount: number
+  ownerNickname: string
+  ownerId: number
+  images: string[]
+  tags: string[]
+  price: number
+  description: string
+  name: string
+  id: number
+  isFavorite?: boolean
+}
